backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2
chromadb>=0.5
pypdf
google-genai>=0.6.0
